#208 [mod] : nested 제거 후 리사이클러뷰 scroll to top 로직 이동
